requirement.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
camelcase==0.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-crispy-forms==1.9.2
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
isort==5.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
pep8==1.7.1
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
